+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+#### Unreleased
+
+* [breaking] - get rid of error-chain. ta::Error -> ta::TaError
+
 #### v0.4.0 - 2020-11-03
 
 * [breaking] Unify parameters for the indicators