+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * [breaking] - get rid of error-chain. ta::Error -> ta::TaError
 * Implement Mean Absolute Deviation (MAD)
 * Implement Commodity Channel Index (CCI)
+* More efficient MoneyFlowIndex
+* Fix identical TypicalPrice bug in MoneyFlowIndex
 
 #### v0.4.0 - 2020-11-03